use memaligned buffers for bio_read
If this is the expected usage I'm not happy about it, because bio_read
accepts a byte offset so who's to say it's reading a whole block
anyway? it works here because (I think) 1024 is a multiple of the
block size.

IMO bio_read should accept unaligned buffers and handle the
requirements of bio_read_block with appropriate use of memcpy. For the
moment, though, this unbreaks everything